+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Current
+- Fix parsing multiple delimeters in the amount, after BigDecimal updates
+
 ## 1.13.0
 - **Breaking change**: check ISO currency code validity when parsing strings with `to_money`
 - Adds `expect_whole_subunits` option when fractional subunits are expected